requirements.txt
mcp>=1.0.0
paramiko>=3.4.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.0
authlib>=1.3.0
pyjwt>=2.8.0